Change atoi to atoll

+++ source/m2mresourceinstance.cpp
@@ -339,11 +339,11 @@ void M2MResourceInstance::get_value(uint8_t *&value, uint32_t &value_length)
 
 int64_t M2MResourceInstance::get_value_int() const
 {
-    int value_int = 0;
+    int64_t value_int = 0;
     uint8_t* buffer = _value;
 
     if(buffer) {
-        value_int = atoi((const char*)buffer);
+        value_int = atoll((const char*)buffer);
     }
     return value_int;
 }
